SPRINT 18. VIRTUALIZATION. DOCKER IMAGE/pom-54-team-12/requirements.txt
asgiref==3.5.2
Django==4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
djangorestframework==3.14.0
django-crispy-forms
django-environ==0.4.5